First commit - experimenting multi drop down
this is one challenging project where I trying to find ways to make the drop down list in pure css, otherwise I go the long route of making javascript. Making the layout and design finished quick. Last on the list is making multi drop down list with css or javascript.